Move PushRegistrationWorker to push package
