+++ app/src/main/kotlin/at/bitfire/davdroid/push/PushRegistrationWorker.kt
@@ -1,4 +1,8 @@
-package at.bitfire.davdroid.syncadapter
+/*
+ * Copyright © All Contributors. See LICENSE and AUTHORS in the root directory for details.
+ */
+
+package at.bitfire.davdroid.push
 
 import android.accounts.Account
 import android.content.Context
